requirements.txt
astroid==2.3.3
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
flask-talisman==0.7.0
Flask-WTF==0.14.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pdfkit==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
flask-mysql==1.5.2
mysql==0.0.2
mysql-connector==2.2.9
flask-mysqldb==0.2.0
pymysql==1.0.2
numpy==1.19.1
